% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ca.R
\name{CA}
\alias{CA}
\title{Simple (efficient) correspondence analysis}
\usage{
CA(obj, nd = NA, suprow = NA, supcol = NA, subsetrow = NA,
  subsetcol = NA, verbose = TRUE)
}
\arguments{
\item{obj}{Matrix with two-way frequency table}

\item{nd}{Number of dimensions to be included in the output; if NA the maximum
possible dimensions are included.}

\item{suprow}{Indices of supplementary rows.}

\item{supcol}{Indices of supplementary columns.}

\item{subsetrow}{Row indices of subset.}

\item{subsetcol}{Column indices of subset.}

\item{verbose}{shows additional ouput about token usage in console}
}
\description{
\code{CA} is a modified version of the \code{ca} function in the
\code{ca} package that consumes less memory in the computation
of simple correspondence analysis.
}
\author{
Michael Greenacre, Oleg Nenadic, Michael Friendly (Modified by Pablo Barbera)
}

