% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tweetscores-package.R
\docType{data}
\name{dict}
\alias{dict}
\title{Dictionary of user IDs and account creation dates}
\usage{
data(dict)
}
\description{
Data frame that contains dates in which a random sample
of 1000 user accounts were created
}
\keyword{datasets}

