% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{estimateDateBreaks}
\alias{estimateDateBreaks}
\title{Estimates dates in which a sample of followers created their accounts}
\usage{
estimateDateBreaks(followers, seed = 777)
}
\arguments{
\item{followers}{List of followers of account}

\item{seed}{Random seed}
}
\description{
Computes dates in which a sample of followers created
their Twitter accounts. This function is then used to estimate
what that list of followers looked like in the past.
}
