% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-ideology.R
\name{estimateIdeology}
\alias{estimateIdeology}
\title{Estimates ideology for a given Twitter user}
\usage{
estimateIdeology(user, friends, verbose = TRUE, method = "MCMC",
  iters = 5000, n.warmup = 1000, thin = 20, ...)
}
\arguments{
\item{user}{screen name of user for which ideology is to be estimated.}

\item{friends}{vector of user IDs that the user for which ideology wants
to be estimated follows. If missing, \code{\link{getFriends}} is called for 
the value of \code{user}.}

\item{verbose}{logical, default is \code{TRUE}, which generates some output
to the R console with information about progress of the sampler.}

\item{method}{"MCMC" (default) for the Metropolis algorithm described in the paper.
"MLE" will use simple maximum likelihood estimation to compute the point estimate.
Note that the standard error for the ML method is likely to be wrong.}

\item{iters}{number of iterations of the metropolis algorithm. Default is 5000}

\item{n.warmup}{warmup period for the sampler. Default is 1000 iterations.}

\item{thin}{thinning of the sampler. Default is 20.}

\item{...}{other options to be passed to the estimation functions}
}
\value{
The function returns a matrix with summary statistics of the posterior
distribution of the two estimated parameters, beta (political interest) and
theta (ideology).
}
\description{
\code{estimateIdeology} estimates ideology for a given user using the
Metropolis algorithm developed in Barbera, 2013. It takes as argument
of the function a list of user IDs indicating who a given user follows.
}
\examples{
\dontrun{
## download list of friends for a given user
 friends <- getFriends(screen_name = "p_barbera", oauth_folder="oauth")
## estimating ideology
 results <- estimateIdeology(friends)
 results['theta', 'mean']
## estimating ideology using ML (fast) method
 results <- estimateIdeology(friends, method="ML")
 results['theta', 'mean']
}

}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}
