% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynamic.R
\name{estimatePastFollowers}
\alias{estimatePastFollowers}
\title{Estimates the follower list of a given user at a point in time
in the past}
\usage{
estimatePastFollowers(followers, breaks, dates_sample, date, verbose = TRUE)
}
\arguments{
\item{followers}{List of followers}

\item{breaks}{user IDs that indicate potential date breaks}

\item{dates_sample}{dates in which the user IDs specified in
\code{breaks} were created}

\item{date}{specific date at which follower list will be estimated}

\item{verbose}{outputs additional information on the console}
}
\description{
Estimates the follower list of a given user at a point in time
in the past
}

