% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-twitter-date.R
\name{formatTwDate}
\alias{formatTwDate}
\title{Converts from Twitter date format to R date format}
\usage{
formatTwDate(datestring, format = "datetime")
}
\arguments{
\item{datestring}{Date string in Twitter format}

\item{format}{Either "date", for Date format; or "datetime" for POSIX.}
}
\author{
Pablo Barbera \email{P.Barbera@lse.ac.uk}
}
