% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dynamic.R
\name{getCreated}
\alias{getCreated}
\title{Returns approximate date in which an account was created based on its
user ID}
\usage{
getCreated(id)
}
\arguments{
\item{id}{Twitter user ID}
}
\description{
\code{getCreated} estimates the date in which an account was created
using a dataset that matches user IDs to dates.
}

