% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-followers.R
\name{getFollowers}
\alias{getFollowers}
\title{Returns the list of user IDs that correspond to a given user's follower}
\usage{
getFollowers(
  screen_name = NULL,
  oauth,
  cursor = -1,
  user_id = NULL,
  verbose = TRUE,
  sleep = 1,
  file = NULL
)
}
\arguments{
\item{screen_name}{user name of the Twitter user for which their followers
will be downloaded}

\item{oauth}{One of the following: either a list with details for an access token
(see example below), a folder where OAuth tokens are stored, or a csv file
with the format: consumer_key, consumer_secret, access_token, access_token_secret.}

\item{cursor}{See \url{https://dev.twitter.com/docs/api/1.1/get/followers/ids}}

\item{user_id}{user id of the Twitter user for which their friends will be
downloaded}

\item{verbose}{If \code{TRUE}, prints information about API calls on console}

\item{sleep}{Number of seconds to sleep between API calls.}

\item{file}{If not \code{NULL}, will store followers list in file instead of in
memory (useful for users with many followers in computer with low RAM memory).}
}
\description{
\code{getFollowers} connects to the REST API of Twitter and returns the
list of followers of a given user. Note that this function allows the
use of multiple OAuth token to make the process more efficient.
}
\examples{
\dontrun{
## Creating OAuth token
 my_oauth <- list(consumer_key = "CONSUMER_KEY",
   consumer_secret = "CONSUMER_SECRET",
   access_token="ACCESS_TOKEN",
   access_token_secret = "ACCESS_TOKEN_SECRET")
## Download list of followers of user "p_barbera"
 followers <- getFollowers(screen_name="p_barbera", oauth=my_oauth)
}

}
