% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-retweets.R
\name{getRetweets}
\alias{getRetweets}
\title{Returns the list of user IDs that retweeted a specific tweet.}
\usage{
getRetweets(id = NULL, oauth, cursor = -1, verbose = TRUE, sleep = 1)
}
\arguments{
\item{id}{The numerical ID of the desired status}

\item{oauth}{One of the following: either a list with details for an access token
(see example below), a folder where OAuth tokens are stored, or a csv file
with the format: consumer_key, consumer_secret, access_token, access_token_secret.}

\item{cursor}{See \url{https://dev.twitter.com/docs/api/1.1/get/statuses/retweeters/ids}}

\item{verbose}{If \code{TRUE}, prints information about API calls on console}

\item{sleep}{Number of seconds to sleep between API calls.}
}
\description{
\code{getRetweets} connects to the REST API of Twitter and returns a
list of up to 100 user IDs belonging to users who have retweeted the
tweet specified by the id parameter.
}
\examples{
\dontrun{
## Creating OAuth token
 my_oauth <- list(consumer_key = "CONSUMER_KEY",
   consumer_secret = "CONSUMER_SECRET",
   access_token="ACCESS_TOKEN",
   access_token_secret = "ACCESS_TOKEN_SECRET")
## Download list of recent user IDs retweeting a tweet by Hillary Clinton
 rts <- getRetweets(id='653733796408377344', oauth=my_oauth)
}

}
