% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-statuses.R
\name{getStatuses}
\alias{getStatuses}
\title{Downloads tweets by their ID from REST API and saves to a json file}
\usage{
getStatuses(
  ids = NULL,
  filename,
  oauth,
  tweet_mode = "extended",
  verbose = TRUE,
  sleep = 1
)
}
\arguments{
\item{ids}{list of tweet IDs to be downloaded}

\item{filename}{Name of file where json tweets will be stored}

\item{oauth}{One of the following: either a list with details for an access token
(see example below), a folder where OAuth tokens are stored, or a csv file
with the format: consumer_key, consumer_secret, access_token, access_token_secret.}

\item{tweet_mode}{if "extended", will return up to 280 characters per tweet.}

\item{verbose}{If \code{TRUE}, prints information about API calls on console}

\item{sleep}{Number of seconds to sleep between API calls.}
}
\description{
Downloads tweets by their ID from REST API and saves to a json file
}
\examples{
\dontrun{
## Creating OAuth token
 my_oauth <- list(consumer_key = "CONSUMER_KEY",
   consumer_secret = "CONSUMER_SECRET",
   access_token="ACCESS_TOKEN",
   access_token_secret = "ACCESS_TOKEN_SECRET")
## Download list of recent user IDs retweeting a tweet by Hillary Clinton
 getStatuses(ids=c('1454115859534950406', '1452687910055002115',
      '1451896893743767555'), filename='example-tweets.json', oauth=my_oauth)
}

}
