% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-users-batch.R
\name{getUsersBatch}
\alias{getUsersBatch}
\title{Returns user data for a vector of Twitter user IDs or screen_names}
\usage{
getUsersBatch(ids = NULL, screen_names = NULL, oauth_folder,
  include_entities = "false", verbose = TRUE, output = NULL)
}
\arguments{
\item{ids}{ids of Twitter users}

\item{screen_names}{user names of the Twitter users}

\item{oauth_folder}{folder where OAuth tokens are stored. It can also be the name
of a file that contains the token, or a csv file with the format: consumer_key,
consumer_secret,access_token,access_token_secret.}

\item{include_entities}{if "true", returned data will include most
recent tweet}

\item{verbose}{shows additional ouput about token usage in console}

\item{output}{If not \code{NULL}, will write user data in raw JSON format
to that file}
}
\description{
\code{getUsersBatch} connects to the REST API of Twitter and returns user
objects (user information) for Twitter users, based on their
screen names or user IDs
}
\examples{
\dontrun{
## Download user data for users "p_barbera" and "barackobama"
 userdata <- getUsersBatch(screen_names=c("p_barbera", "BarackObama"),
   oauth_folder="~/Dropbox/credentials")
}

}
\author{
Pablo Barbera \email{P.Barbera@lse.ac.uk}
}
