% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetscores-package.R
\docType{data}
\name{refdata}
\alias{refdata}
\title{Summary ideology estimates for elites}
\usage{
data(refdata)
}
\description{
Data frame that contains ideal points for elites
}
\keyword{datasets}

