% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetscores-package.R
\docType{data}
\name{refdataCA}
\alias{refdataCA}
\title{Column coordinates from correspondence analysis}
\usage{
data(refdataCA)
}
\description{
List that contains output from correspondence analysis model in
Barbera et al, 2015, Psychological Science
}
\keyword{datasets}
