% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scrape-congress.R
\name{scrapeCongressData}
\alias{scrapeCongressData}
\title{Scrape the list of Twitter accounts for Members of US Congress from}
\usage{
scrapeCongressData()
}
\description{
Scrape the list of Twitter accounts for Members of US Congress from
}
\examples{
\dontrun{
 congress <- scrapeCongressData()
}
}

