% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-users.R
\name{searchUsers}
\alias{searchUsers}
\title{Returns list of users related to a search query}
\usage{
searchUsers(q = NULL, count = 100, oauth, verbose = TRUE)
}
\arguments{
\item{q}{The search query to run against people search}

\item{count}{Number of potential user results to retrieve}

\item{oauth}{One of the following: either a list with details for an access token
(see example below), a folder where OAuth tokens are stored, or a csv file
with the format: consumer_key, consumer_secret, access_token, access_token_secret.}

\item{verbose}{shows additional ouput about token usage in console}
}
\description{
\code{searchUsers} Provides a simple, relevance-based search interface
to public user accounts on Twitter. Try querying by topical interest,
full name, company name, location, or other criteria. Exact match searches
are not supported. Only the first 1,000 matches are available.
}
\examples{
\dontrun{
## Search users using query "pablo"
 userdata <- searchUsers(q="pablo", count=200, oauth=my_oauth)
}

}
