% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplementary.R
\name{supplementaryColumns}
\alias{supplementaryColumns}
\title{Projects additional columns (political accounts) to a latent ideological space
using correspondence analysis}
\usage{
supplementaryColumns(res, points)
}
\arguments{
\item{res}{Output from \code{CA} function}

\item{points}{Boolean vector that indicates whether a political account is
followed by the users included in the full model}
}
\description{
\code{supplementaryColumns} takes additional columns of a follower matrix
and projects them to the latent ideological space using the parameters
of an already-fitted correspondence analysis model.
Code was adapted from the \code{ca} function in the \code{ca} package
}
\author{
Michael Greenacre, Oleg Nenadic, Michael Friendly (Modified by Pablo Barbera)
}
