% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/supplementary.R
\name{supplementaryRows}
\alias{supplementaryRows}
\title{Projects additional rows (users) to a latent ideological space
using correspondence analysis}
\usage{
supplementaryRows(res, points)
}
\arguments{
\item{res}{Output from \code{CA} function}

\item{points}{Boolean vector that indicates whether a user follows the
political accounts used to estimate the full model}
}
\description{
\code{supplementaryRows} takes additional rows of a follower matrix
and projects them to the latent ideological space using the parameters
of an already-fitted correspondence analysis model.
Code was adapted from the \code{ca} function in the \code{ca} package
}
\author{
Michael Greenacre, Oleg Nenadic, Michael Friendly (Modified by Pablo Barbera)
}

