% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twideology-functions.R
\name{tracePlot}
\alias{tracePlot}
\title{Displays trace plots of MCMC chains}
\usage{
tracePlot(results, par = "theta")
}
\arguments{
\item{results}{object of class 'twideology'}

\item{par}{parameter for which trace plot is to be displayed}
}
\description{
Displays trace plots of MCMC chains
}
\examples{
\dontrun{
## Creating OAuth token
 my_oauth <- list(consumer_key = "CONSUMER_KEY",
   consumer_secret = "CONSUMER_SECRET",
   access_token="ACCESS_TOKEN",
   access_token_secret = "ACCESS_TOKEN_SECRET")
## download list of friends for a given user
 friends <- getFriends(screen_name = "p_barbera", oauth=my_oauth)
## estimating ideology
 results <- estimateIdeology(user, friends)
## summary
 summary(results)
## trace plot
 tracePlot(results, "theta")
## ideology plot
 plot(results)
}

}
