% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-list.R
\name{getList}
\alias{getList}
\title{Returns the list of users added to a Twitter list}
\usage{
getList(list_id, oauth, cursor = -1)
}
\arguments{
\item{list_id}{ID of the list}

\item{oauth}{One of the following: either a list with details for an access token
(see example below), a folder where OAuth tokens are stored, or a csv file
with the format: consumer_key, consumer_secret, access_token, access_token_secret.}

\item{cursor}{See \url{https://dev.twitter.com/docs/api/1.1/get/lists/members}}
}
\description{
\code{getList} connects to the REST API of Twitter and returns a
data frame with information about users included in a Twitter list
}
\examples{
\dontrun{
## Creating OAuth token
 my_oauth <- list(consumer_key = "CONSUMER_KEY",
   consumer_secret = "CONSUMER_SECRET",
   access_token="ACCESS_TOKEN",
   access_token_secret = "ACCESS_TOKEN_SECRET")
## Download Twitter list of U.S. Governors created by @cspan
 accts <- getList(list_id="7560205", oauth=my_oauth)
}

}
