% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-users-batch.R
\name{getUsersBatch}
\alias{getUsersBatch}
\title{Returns user data for a vector of Twitter user IDs or screen_names}
\usage{
getUsersBatch(
  ids = NULL,
  screen_names = NULL,
  oauth,
  include_entities = "false",
  verbose = TRUE,
  output = NULL
)
}
\arguments{
\item{ids}{ids of Twitter users}

\item{screen_names}{user names of the Twitter users}

\item{oauth}{One of the following: either a list with details for an access token
(see example below), a folder where OAuth tokens are stored, or a csv file
with the format: consumer_key, consumer_secret, access_token, access_token_secret.}

\item{include_entities}{if "true", returned data will include most
recent tweet}

\item{verbose}{shows additional ouput about token usage in console}

\item{output}{If not \code{NULL}, will write user data in raw JSON format
to that file}
}
\description{
\code{getUsersBatch} connects to the REST API of Twitter and returns user
objects (user information) for Twitter users, based on their
screen names or user IDs
}
\examples{
\dontrun{
## Creating OAuth token
 my_oauth <- list(consumer_key = "CONSUMER_KEY",
   consumer_secret = "CONSUMER_SECRET",
   access_token="ACCESS_TOKEN",
   access_token_secret = "ACCESS_TOKEN_SECRET")
## Download user data for users "p_barbera" and "barackobama"
 userdata <- getUsersBatch(screen_names=c("p_barbera", "BarackObama"),
   oauth=my_oauth)
}

}
