% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twideology-functions.R
\name{plot.twideology}
\alias{plot.twideology}
\title{Displays estimated ideology with other reference ideal points}
\usage{
\method{plot}{twideology}(x, ...)
}
\arguments{
\item{x}{object of class 'twideology'}

\item{...}{ignored}
}
\examples{
\dontrun{
## Creating OAuth token
 my_oauth <- list(consumer_key = "CONSUMER_KEY",
   consumer_secret = "CONSUMER_SECRET",
   access_token="ACCESS_TOKEN",
   access_token_secret = "ACCESS_TOKEN_SECRET")
## download list of friends for a given user
 friends <- getFriends(screen_name = "p_barbera", oauth=my_oauth)
## estimating ideology
 results <- estimateIdeology(user, friends)
## trace plot
 tracePlot(results, "theta")
## ideology plot
 plot(results)
}

}
\author{
Pablo Barbera \email{P.Barbera@lse.ac.uk}
}
