% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-congress.R
\name{scrapeCongressData}
\alias{scrapeCongressData}
\title{Scrape the list of Twitter accounts for Members of US Congress from
the unitedstates GitHub account}
\usage{
scrapeCongressData(commit = "master")
}
\arguments{
\item{commit}{Commit from which data will be pulled}
}
\examples{
\dontrun{
 congress <- scrapeCongressData()
}

}
