% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape-congress.R
\name{scrapeCongressData}
\alias{scrapeCongressData}
\title{Scrape the social media accounts of Members of US Congress}
\usage{
scrapeCongressData(commit = "master")
}
\arguments{
\item{commit}{Commit from which data will be pulled}
}
\description{
\code{scrapeCongressData} downloads the list of social media accounts
(Twitter and Facebook) for the Members of the US Congress from the
unitedstates GitHub account at any point in time, and returns a data
frame.
}
\examples{
\dontrun{
 congress <- scrapeCongressData()
}

}
